refactor(ui): Improve character creation components and string resources
- Split UI components into separate files
- Move components to proper packages (screens, components)
- Extract string resources for internationalization
- Improve code organization in CharacterCreationActivity
- Add proper dependency initialization
- Enhance component reusability and maintainability
