+++ app/src/main/java/com/github/cstrerath/uncover/ui/screens/character/CreateCharacterButton.kt
@@ -1,0 +1,23 @@
+package com.github.cstrerath.uncover.ui.screens.character
+
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.material3.Button
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.stringResource
+import com.github.cstrerath.uncover.R
+
+@Composable
+internal fun CreateCharacterButton(
+    enabled: Boolean,
+    onClick: () -> Unit
+) {
+    Button(
+        onClick = onClick,
+        enabled = enabled,
+        modifier = Modifier.fillMaxWidth()
+    ) {
+        Text(stringResource(R.string.create_character))
+    }
+}
